+++ home-helper-front/src/components/page/exchange_item/ExchangeItemRegisterPage.tsx
@@ -1,6 +1,6 @@
 import { FC } from "react";
 import { HomeHelperLayout } from "@components/ui/Layout/HomeHelperLayout";
-import { ExchangeItemRegisterContainer } from "@components/container/exchange_item/register/ExchangeItemRegisterContainer";
+import { ExchangeItemRegisterContainer } from "@components/presenter/exchange_item/register/ExchangeItemRegisterContainer";
 
 export const Page: FC = () => (
   <HomeHelperLayout>
