+++ home-helper-front/src/components/page/help_item/HelpItemRegisterPage.tsx
@@ -1,6 +1,6 @@
 import { FC } from "react";
 import { HomeHelperLayout } from "@components/ui/Layout/HomeHelperLayout";
-import { HelpItemRegisterContainer } from "@components/container/help_item/register/HelpItemRegisterContainer";
+import { HelpItemRegisterContainer } from "@components/presenter/help_item/register/HelpItemRegisterContainer";
 
 export const Page: FC = () => (
   <HomeHelperLayout>
