+++ home-helper-front/src/components/page/point_history/PointHistoryPage.tsx
@@ -1,5 +1,5 @@
 import { FC } from "react";
-import { PointHistoryContainer } from "@components/container/point_history/PointHistoryContainer";
+import { PointHistoryContainer } from "@components/presenter/point_history/PointHistoryContainer";
 import { HomeHelperLayout } from "@components/ui/Layout/HomeHelperLayout";
 
 export const Page: FC = () => (
