+++ home-helper-front/src/components/page/top/TopPage.tsx
@@ -1,5 +1,5 @@
 import { FC } from "react";
-import { TopContainer } from "@components/container/top/TopContainer";
+import { TopContainer } from "@components/presenter/top/TopContainer";
 import { TopLayout } from "@components/ui/Layout/TopLayout";
 
 export const Page: FC = () => (
