+++ home-helper-front/src/components/presenter/exchange_item/register/ExchangeItemRegisterContainer.tsx
@@ -1,18 +1,21 @@
-import { HelpItemRegisterPresenter } from "@components/presenter/help_item/register/HelpItemRegisterPresenter";
-import { useState } from "react";
-import { registerHelpItem } from "@hooks/help_item/registerHelpItem";
-import { showNotification } from "@hooks/notification/showNotification";
+/*
+ * Copyright (c) 2023 Ryo-Kgym.
+ */
 
-export const HelpItemRegisterContainer = () => {
+import { useState } from "react";
+import { showNotification } from "@hooks/notification/showNotification";
+import { ExchangeItemRegisterPresenter } from "./ExchangeItemRegisterPresenter";
+import { registerExchangeItem } from "@hooks/exchange_item/registerExchangeItem";
+
+export const ExchangeItemRegisterContainer = () => {
   const [itemName, setItemName] = useState<string>("");
   const [point, setPoint] = useState<number>(0);
-  const [memo, setMemo] = useState<string>("");
 
   const register = () => {
     if (itemName.length < 1) {
       return;
     }
-    registerHelpItem(itemName, point, memo).then((r) => {
+    registerExchangeItem(itemName, point).then((r) => {
       showNotification({
         title: "登録完了",
         message: r[0],
@@ -21,13 +24,11 @@ export const HelpItemRegisterContainer = () => {
   };
 
   return (
-    <HelpItemRegisterPresenter
+    <ExchangeItemRegisterPresenter
       itemName={itemName}
       setItemName={setItemName}
       point={point}
       setPoint={setPoint}
-      memo={memo}
-      setMemo={setMemo}
       registerOnClick={register}
     />
   );
