+++ home-helper-front/src/components/presenter/house_hold/HouseHoldContainer.tsx
@@ -1,5 +1,5 @@
 import { FC } from "react";
-import { HouseHoldPresenter } from "@components/presenter/house_hold/HouseHoldPresenter";
+import { HouseHoldPresenter } from "./HouseHoldPresenter";
 
 export const HouseHoldContainer: FC = () => {
   return <HouseHoldPresenter />;
